Allow duration/speed to be retained on static animations.
Previously, static animations had `Duration 0` so they played nicely
with the lifecycle functions. The desired behavior is that static
animations are always done, never scheduled or running, but before
they were both done and scheduled. Now, we handle that case explictly
and correctly. Furthermore, setting the to or retargetting a static
animation will now use the default duration unless one has been set,
in which case it will use that one.